requirements.txt
accelerate==1.10.1
numpy==1.26.4
pandas==2.3.3
plotly==6.5.0
python-dotenv==1.2.1
requests==2.32.5
streamlit==1.50.0
tokenizers==0.22.1
torch==2.2.2
tqdm==4.67.1
transformers==4.57.3

openpyxl==3.1.5

SQLAlchemy==1.4.41
psycopg2-binary==2.9.11
alembic==1.9.2
PyJWT==2.10.1
cryptography==41.0.7
matplotlib==3.9.4
pydantic==2.12.5
pytest==7.4.4
